refactor: extract token interaction logic into trait
